Fix for most raytracing failuers
The function func_within_plasma didn't cover some cases
leading to multiple crashes. This is fixed now.

Realized that the derivatives of the hamiltonian
are incorrect if X, Y or N_par are zero.

While these cases are unlikely they were not raising
an exception.

Fixed the binning of the BPD crashing the program
if a ray does not pass through the plasma,

Added F2PY debugging to the makefile.